リファクタリング
誤った依存関係と例外処理を整理
