+++ src/main/java/jp/co/tis/gsp/tools/dba/dialect/Dialect.java
@@ -17,50 +17,34 @@
 package jp.co.tis.gsp.tools.dba.dialect;
 
 import java.io.File;
-import java.io.FileNotFoundException;
-import java.io.FileOutputStream;
 import java.io.IOException;
-import java.io.OutputStream;
 import java.nio.charset.Charset;
 import java.sql.Connection;
 import java.sql.DatabaseMetaData;
-import java.sql.Date;
-import java.sql.DriverManager;
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
-import java.sql.ResultSetMetaData;
 import java.sql.SQLException;
 import java.sql.Statement;
-import java.sql.Timestamp;
 import java.sql.Types;
-import java.text.SimpleDateFormat;
-import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
-import java.util.Map;
-import java.util.Set;
-import java.util.TreeMap;
 
 import javax.persistence.GenerationType;
 
 import org.apache.commons.io.FileUtils;
 import org.apache.maven.plugin.MojoExecutionException;
 import org.seasar.extension.jdbc.gen.meta.DbTableMeta;
-import org.seasar.extension.jdbc.util.ConnectionUtil;
 import org.seasar.framework.util.DriverManagerUtil;
-import org.seasar.framework.util.OutputStreamUtil;
 import org.seasar.framework.util.ResultSetUtil;
 import org.seasar.framework.util.StatementUtil;
 import org.seasar.framework.util.StringUtil;
-
-import com.csvreader.CsvWriter;
 
 import jp.co.tis.gsp.tools.db.AlternativeGenerator;
 import jp.co.tis.gsp.tools.db.TypeMapper;
 import jp.co.tis.gsp.tools.dba.dialect.param.ExportParams;
 import jp.co.tis.gsp.tools.dba.dialect.param.ImportParams;
-import jp.co.tis.gsp.tools.dba.util.CsvDataLoader;
-import jp.co.tis.gsp.tools.dba.util.DialectUtil;
+import jp.co.tis.gsp.tools.dba.util.CsvExporter;
+import jp.co.tis.gsp.tools.dba.util.CsvLoader;
 import jp.co.tis.gsp.tools.dba.util.SqlExecutor;
 
 public abstract class Dialect {
@@ -72,8 +56,6 @@ public abstract class Dialect {
 	protected DatabaseMetaData metaData = null;
 	protected static final int UN_USABLE_TYPE = -999;
 	
-    protected final SimpleDateFormat sdfDate = new SimpleDateFormat("yyyy-MM-dd");
-    protected final SimpleDateFormat sdfTimestamp = new SimpleDateFormat("yyyy-MM-dd HH:mm:ss");
 
     final Charset UTF8 = Charset.forName("UTF-8");
     
@@ -83,24 +65,23 @@ public abstract class Dialect {
     protected final String DATA_DIR_NAME = "dataDirectory";
     
 	public void exportSchema(String user, String password, String schema, ExportParams params) throws MojoExecutionException {
-	    
-	    // 汎用タイプのエクスポート処理
-	    exportSchemaGeneral(user, password, schema, params);
-	}
-	
-	protected void exportSchemaGeneral(String user, String password, String schema, ExportParams params) throws MojoExecutionException {
-	       // CSVデータ出力
-        exportCsv(user, password, schema, new File(params.getOutputDirectory(), DATA_DIR_NAME), UTF8);
+       // CSVデータ出力
+	    CsvExporter exporter = new CsvExporter(url, driver, schema, user, password,  new File(params.getOutputDirectory(), DATA_DIR_NAME), UTF8);
+	    try {
+            exporter.execute();
+        } catch (Exception e1) {
+            new MojoExecutionException("CSVデータ出力処理でエラーが発生しました:", e1);
+        }
         
         // DDL & extraDDLの収集
         try {
             FileUtils.copyDirectory(params.getDdlDirectory(), new File(params.getOutputDirectory(), DDL_DIR_NAME));
             FileUtils.copyDirectory(params.getExtraDdlDirectory(), new File(params.getOutputDirectory(), EXTRADDL_DIR_NAME));
         } catch (IOException e) {
-            throw new MojoExecutionException("DDLとデータファイルのコピーに失敗しました。", e);
+            throw new MojoExecutionException("DDLとデータファイルのコピーに失敗しました:", e);
         }
 	}
-
+	
 	/**
 	 * 指定したスキーマ内の全テーブル・ビュー・シーケンスを削除します。
 	 * 
@@ -118,24 +99,27 @@ public abstract class Dialect {
 
 	public void importSchema(String user, String password, String schema, ImportParams params) throws MojoExecutionException {
 	    
-	    // 汎用タイプのエクスポートデータのインポート
-	    importSchemaGeneral(user, password, schema, params);
+	    File inputDir = params.getInputDirectory();
+	    
+        // DDLの実行
+        SqlExecutor sqlExecutor = new SqlExecutor(schema, user, password, new File(inputDir, DDL_DIR_NAME), new File(inputDir, EXTRADDL_DIR_NAME), 
+                params.getDelimiter(), params.getOnError(), params.getLogger());
+        try {
+            sqlExecutor.execute();
+        } catch (SQLException e) {
+            throw new MojoExecutionException("DDLの実行に失敗しました:", e);
+        }
+
+        // LoadDataの実行
+        CsvLoader dataLoader = new CsvLoader(url, driver, schema, user, password, new File(inputDir, DATA_DIR_NAME), UTF8,
+                null, params.getOnError(), params.getLogger());
+        try {
+            dataLoader.execute();
+        } catch (Exception e) {
+            throw new MojoExecutionException("CSVデータのロード処理で失敗しました:", e);
+        }
 	}
 
-	protected void importSchemaGeneral(String user, String password, String schema, ImportParams params) throws MojoExecutionException {
-	       File inputDir = params.getInputDirectory();
-	        
-	        // DDLの実行
-	        SqlExecutor sqlExecutor = new SqlExecutor(schema, user, password, new File(inputDir, DDL_DIR_NAME), new File(inputDir, EXTRADDL_DIR_NAME), 
-	                params.getDelimiter(), params.getOnError(), params.getLogger());
-	        sqlExecutor.execute();
-	        
-	        // LoadDataの実行
-	        CsvDataLoader dataLoader = new CsvDataLoader(url, driver, schema, user, password, new File(inputDir, DATA_DIR_NAME), UTF8,
-	                null, params.getOnError(), params.getLogger());
-	        dataLoader.execute();
-	}
-	    
 	/**
 	 * ユーザを作成します。
 	 * 
@@ -371,198 +355,4 @@ public abstract class Dialect {
             StatementUtil.close(stmt);
         }
     }
-    
-    /**
-     * 指定スキーマのテーブルデータをcsvで出力します。
-     * 
-     * @param adminUser
-     * @param adminPassword
-     * @param schema
-     * @param outputDir
-     * @param charset
-     * @throws MojoExecutionException
-     */
-    public void exportCsv(String adminUser, String adminPassword, String schema, File outputDir, Charset charset) 
-            throws MojoExecutionException {
-        DriverManagerUtil.registerDriver(driver);
-        Connection conn = null;
-        try {
-            conn = DriverManager.getConnection(url, adminUser, adminPassword);
-            String normalizeSchemaName = normalizeSchemaName(schema);
-        
-            final List<String> tableList = getTableNames(conn.getMetaData(), normalizeSchemaName,
-                    new String[] { "TABLE" });
-
-            for (String tableName : tableList) {
-                writeToCSV(conn, normalizeSchemaName, tableName, new File(outputDir, tableName + ".csv"), charset);
-            }
-
-        } catch (SQLException e) {
-            throw new MojoExecutionException("DBに接続できませんでした。driverおよびurlの設定が正しいか確認してください", e);
-        } catch (FileNotFoundException e) {
-            throw new MojoExecutionException("CSVファイルの出力中にエラーが発生しました。", e);
-        } catch (IOException e) {
-            throw new MojoExecutionException("CSVファイルの出力中にエラーが発生しました。", e);
-        } finally {
-            ConnectionUtil.close(conn);
-        }
-    }
-
-    /**
-     * 指定スキーマに属するテーブル名のリストを取得します。
-     * 
-     * @param metaData
-     * @param normalizedSchemaName
-     * @param types
-     * @return
-     * @throws SQLException
-     */
-    public List<String> getTableNames(DatabaseMetaData metaData, String normalizedSchemaName, String[] types)
-            throws SQLException {
-        List<String> allNames = new ArrayList<String>();
-        ResultSet rs = null;
-        try {
-            rs = metaData.getTables(null, normalizedSchemaName, "%", types);
-            while (rs.next()) {
-                allNames.add(rs.getString("TABLE_NAME"));
-            }
-        } finally {
-            if (rs != null) {
-                rs.close();
-            }
-        }
-        return allNames;
-    }
-
-    /**
-     * SELECT文の結果をCSVファイルにカラム名ヘッダー付きで出力します。
-     * 
-     * @param metaData
-     * @param resultSet
-     * @param csvFile
-     * @param charset
-     * @throws SQLException
-     * @throws IOException
-     */
-    public void writeToCSV(Connection conn, String schema, String tableName, File csvFile, Charset charset)
-            throws SQLException, IOException {
-
-        Dialect dialect = DialectUtil.getDialect();
-        Statement stmt = null;
-        ResultSet resultSet = null;
-
-        OutputStream out = new FileOutputStream(csvFile);
-        String[] columnHeaders;
-        String[] values = null;
-        CsvWriter csvWriter = null;
-
-        try {
-            stmt = conn.createStatement();
-            resultSet = stmt
-                    .executeQuery("SELECT * FROM " + tableName + getOrderString(conn.getMetaData(), schema, tableName));
-            ResultSetMetaData meta = resultSet.getMetaData();
-
-            csvWriter = new CsvWriter(out, ',', charset);
-            csvWriter.setForceQualifier(true);
-            csvWriter.setRecordDelimiter('\n');
-
-            // カラムヘッダーの出力
-            int columnCount = meta.getColumnCount();
-            columnHeaders = new String[columnCount];
-            for (int i = 1; i <= columnCount; i++) {
-                columnHeaders[i - 1] = meta.getColumnName(i);
-            }
-            csvWriter.writeRecord(columnHeaders);
-
-            // データの出力
-            while (resultSet.next()) {
-                values = new String[columnCount];
-                for (int i = 1; i <= columnCount; i++) {
-                    resultSet.getString(i);
-                    int sqlType = dialect.guessType(conn, schema, tableName, meta.getColumnName(i));
-                    values[i - 1] = convert(resultSet, i, sqlType);
-                }
-                csvWriter.writeRecord(values);
-            }
-        } finally {
-            csvWriter.flush();
-            csvWriter.close();
-
-            ResultSetUtil.close(resultSet);
-            StatementUtil.close(stmt);
-            OutputStreamUtil.close(out);
-        }
-    }
-
-    /**
-     * SQLのタイプに応じて値の書式を変換します。
-     * 
-     * @param resultSet
-     * @param meta
-     * @param index
-     * @return
-     * @throws SQLException
-     */
-    protected String convert(ResultSet resultSet, int index, int sqlType) throws SQLException {
-
-        String value = "";
-        switch (sqlType) {
-        case Types.DATE:
-            Date date = resultSet.getDate(index);
-            if (date != null) {
-                value = sdfDate.format(date);
-            }
-            break;
-        case Types.TIMESTAMP:
-            Timestamp timestamp = resultSet.getTimestamp(index);
-            if (timestamp != null) {
-                value = sdfTimestamp.format(timestamp);
-            }
-            break;
-        default:
-            value = resultSet.getString(index);
-            break;
-        }
-
-        return value;
-    }
-
-    /**
-     * 主キー項目でソートをかけるorderBy文字列を生成します。
-     * 
-     * @param meta
-     * @param tableName
-     * @return
-     * @throws SQLException
-     */
-    protected String getOrderString(DatabaseMetaData meta, String schema, String tableName) throws SQLException {
-        String orderByString = "";
-        ResultSet rs = null;
-        Map<Short, String> keyMap = new TreeMap<Short, String>();
-
-        try {
-            rs = meta.getPrimaryKeys(null, schema, tableName);
-            while (rs.next()) {
-                keyMap.put(rs.getShort("KEY_SEQ"), rs.getString("COLUMN_NAME"));
-            }
-        } finally {
-            if (rs != null) {
-                rs.close();
-            }
-        }
-
-        Set<Short> keySet = keyMap.keySet();
-
-        if (keySet.size() > 0) {
-            orderByString = " ORDER BY";
-            for (Short key : keySet) {
-                String preStr = key.shortValue() > 1 ? "," : " ";
-                orderByString += (preStr + keyMap.get(key));
-            }
-        }
-
-        return orderByString;
-    }
-
-
 }
